read hyperparams from files

+++ run_score_report.py
@@ -1,5 +1,6 @@
 import datetime
 import pandas as pd
+import os.path
 
 from features.data_provider import all_features, other_features, player_features, DataLoader
 from models.helpers import get_default_parameters
@@ -19,26 +20,32 @@ tournament_parameters = [
     ('data/original/wc_2010_games_real.csv', 'data/original/wc_2010_bets.csv', "2010-06-11")
 ]
 feature_sets = [
-    ("all_features", all_features),
-    ("general_features", other_features),
-    ("player_features", player_features)
+    ("all_features", all_features, "score_hyperparam_optimization_all_features.csv"),
+    ("general_features", other_features, "score_hyperparam_optimization_general_features.csv"),
+    ("player_features", player_features, "score_hyperparam_optimization_player_features.csv")
 ]
 
 file_name = "score_report_full.txt"
 
 reports = []
-for (name, feature_set) in feature_sets:
+for (name, feature_set, fname) in feature_sets:
     write_log(file_name, str(datetime.datetime.now()))
     write_log(file_name, f"Running test for feature set: {name}", print_text=True)
 
     data_loader = DataLoader(feature_set)
-    Xhome, yhome, Xaway, yaway = data_loader.get_all_data(["home_score", "away_score"])
-    _, outcomes = data_loader.get_all_data("home_win")
 
-    params = get_default_parameters()
-    arguments = get_cv_grid_search_arguments(params, Xhome)
-    results = run_grid_search_for_score(arguments, Xhome, yhome, Xaway, yaway, outcomes)
-    results.to_csv(f"score_hyperparam_optimization_{name}.csv")
+    if os.path.isfile(fname):
+        write_log(file_name, f"Hyperparameters found for: {name}", print_text=True)
+        results = pd.read_csv(fname)
+    else:
+        Xhome, yhome, Xaway, yaway = data_loader.get_all_data(["home_score", "away_score"])
+        _, outcomes = data_loader.get_all_data("home_win")
+
+        params = get_default_parameters()
+        arguments = get_cv_grid_search_arguments(params, Xhome)
+        results = run_grid_search_for_score(arguments, Xhome, yhome, Xaway, yaway, outcomes)
+        results.to_csv(f"score_hyperparam_optimization_{name}.csv")
+
     best_params = results.sort_values(['test_acc', 'test_logloss'], ascending=[False, True]).iloc[0]
     best_params_dict = best_params.to_dict()
     write_log(file_name, str(best_params_dict), print_text=True)
